% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalOneRun.R
\name{finalOneRun}
\alias{finalOneRun}
\title{Building local models in parallel for all the time points}
\usage{
finalOneRun(y, X, c_init, zeta_init, nusq_init, w_init, v0, a1, a2, q, Nmcmc,
  ind)
}
\arguments{
\item{y}{binary response correspoding to \eqn{n} subjects}

\item{X}{\eqn{n} x \eqn{p} x \eqn{tau} - tensor (EEG data) for \eqn{n} subjects, \eqn{p} locations and \eqn{tau} time points}

\item{c_init}{initial value for \eqn{c}}

\item{zeta_init}{initial values for \eqn{\zeta}}

\item{nusq_init}{initial values for \eqn{\nu^2}}

\item{w_init}{initial value for \eqn{w} - the complexity parameter}

\item{v0}{hyperparameter \eqn{v0}}

\item{a1}{hyperparameter \eqn{a1}}

\item{a2}{hyperparameter \eqn{a2}}

\item{q}{hyperparameter \eqn{q}}

\item{Nmcmc}{number of MCMC samples to generate}

\item{ind}{indices of MCMC samples to use after burnin ad thinning}
}
\description{
Builds local models in parallel by detecting the number of cores (\eqn{k}) on the machine. This function builds models at \eqn{tau} time points parllely using all avialable cores.
}
\examples{
finalOneRun()
}
\keyword{finalOneRun()}
