% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapEst.R
\name{mapEst}
\alias{mapEst}
\title{Compute MAP estimate}
\usage{
mapEst(chain, burnin = 0, thin = 1)
}
\arguments{
\item{chain}{sequence of MCMC samples}

\item{burnin}{Number of initial samples to drop}

\item{thin}{Number of samples to be dropped between two consecutive selections in the chain}
}
\description{
Given a sequence of MCMC samples this function computes the MAP estimate of the posterior distribution.
}
\examples{
mapEst()
}
\keyword{mapEst()}
