% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n0beta.R
\name{n0beta}
\alias{n0beta}
\title{First-stage variable selection}
\usage{
n0beta(beta, method, w = NULL, const = NULL)
}
\arguments{
\item{beta}{\eqn{tau} x \eqn{p} matrix of estimates of beta}

\item{method}{thresholding method to be used 'Zcut', 'Wcut' or 'Ccut'}

\item{w}{estimates of complexity parameters required if method == 'Wcut'}

\item{const}{threshold required if method == 'Ccut'}
}
\description{
First-stage variable selection using the 'MCMC estimates of local Bayesian modeling 'Zcut', 'Wcut' or 'Ccut' method
}
\examples{
n0beta()
}
\keyword{n0beta()}
