% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Prediction for subjects}
\usage{
pred(n0b, bInd, X.test, cEst)
}
\arguments{
\item{n0b}{\eqn{tau} x \eqn{p} matrix of estimates of beta as returned by varSelc function}

\item{bInd}{indices of selected locations after two-stage variable selection as returned by varSelc function}

\item{X.test}{test data with dimensions \eqn{m} x \eqn{p} x \eqn{tau}}

\item{cEst}{\eqn{tau} estimates of \eqn{c} for each local models}
}
\description{
Prediction of the alcoholic status using the estimates from local Bayesian modeling and variable selection
}
\examples{
pred()
}
\keyword{pred()}
