% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelc.R
\name{varSelc}
\alias{varSelc}
\title{Performing two-stage variable selection}
\usage{
varSelc(b, method = "Zcut", wEst = NULL, bf.thres)
}
\arguments{
\item{b}{\eqn{tau} x \eqn{p} matrix of estimates of beta}

\item{method}{thresholding method to be used 'Zcut' or 'Wcut', defaults to 'Zcut'}

\item{wEst}{estimates of complexity parameters required if method == 'Wcut'}

\item{bf.thres}{threshold to be used to compare Bayes factor}
}
\description{
Performing two-stage variable selection using the MCMC estimates of local Bayesian modeling
}
\examples{
varSelc()
}
\keyword{varSelc()}
